POC: mistral-7B inference and foundational game mechanics
This commit introduces the Inference.ipynb notebook in which a quantized
version of mistral-7B is used for inference. Furthermore, the basic flow
of the text-based adventure game has also been introduced in which the
prompt has been set to force the LLM to generate the story and the
choices as a JSON object. The response is then parsed and the story
along with the selected choice is added back to the prompt.

Signed-off-by: Tariq Mehmood <tariq9mehmood9@gmail.com>