fine-tuning: initial attempts
fine-tuning-v1 notebook makes use of output_data.json for fine-tuning
but we observed the the model was generating the [INST] [/INST] tags as
well and there was a lot of repetition.

fine-tuning-v2 is an improvement over v1 after we identified the root
cause of INST tag generation by the model which was happening because
the EOS tag </s> was not appropriately placed while preparing the
dataset. So now instead of preparing the dataset manually, we utilized
the apply_chat_template from the tokenizer library on our dataset having
two roles i.e., Assistant and User. v2 is finetuned using data.jsonl.
While it shows better results compared to v1, they are still far from
the desired output as the model is still generating repetition.

dataset: https://huggingface.co/datasets/PocketDoc/Floyd-Text-Adventures

Signed-off-by: Tariq Mehmood <tariq9mehmood9@gmail.com>