Inference: integrate all the bits and pieces into one file
This file now uses two models, one has been finetuned on the stories
dataset and we use it to generate narratives and the other one is the
base mistral-7b-instruct (quantized) model which we use for generating
choices for a given narrative as well as to process rag queries.

We use bert summarizer to summarize the current narrative and create an
image generation prompt for dall-e.

Signed-off-by: Tariq Mehmood <tariq9mehmood9@gmail.com>